VmID and TxID move to simulator_context
