+++ src/simulator_context.rs
@@ -1,4 +1,3 @@
-use crate::global_data::{TxID, VmID};
 use std::{
     cell::RefCell,
     collections::HashMap,
@@ -336,5 +335,44 @@ impl Fd {
     }
     pub fn is_read(&self) -> bool {
         self.0 % 2 == 0
+    }
+}
+
+#[derive(Default, PartialEq, Eq, Clone, Hash, Debug)]
+pub struct TxID(u64);
+impl From<u64> for TxID {
+    fn from(value: u64) -> Self {
+        Self(value)
+    }
+}
+impl From<TxID> for u64 {
+    fn from(value: TxID) -> Self {
+        value.0
+    }
+}
+impl TxID {
+    pub fn next(&mut self) -> Self {
+        self.0 += 1;
+        self.clone()
+    }
+}
+
+#[derive(Default, PartialEq, Eq, Clone, Hash, Debug)]
+pub struct VmID(u64);
+impl From<u64> for VmID {
+    fn from(value: u64) -> Self {
+        Self(value)
+    }
+}
+impl From<VmID> for u64 {
+    fn from(value: VmID) -> Self {
+        value.0
+    }
+}
+impl VmID {
+    pub fn next(&mut self) -> Self {
+        let id = self.clone();
+        self.0 += 1;
+        id
     }
 }
